DireccEmbarque_Cobrador

+++ StatusERP.Services/Implementations/AS/CobradorService.cs
@@ -1,14 +1,11 @@
 ﻿using Microsoft.Extensions.Logging;
 using StatusERP.DataAccess.Repositories.AS;
+using StatusERP.DataAccess.Repositories.ERPADMIN.Interfaces;
 using StatusERP.Dto.Request.AS;
 using StatusERP.Dto.Response;
+using StatusERP.Entities;
 using StatusERP.Entities.AS.Tablas;
 using StatusERP.Services.Interfaces.AS;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace StatusERP.Services.Implementations.AS
 {
@@ -16,54 +13,77 @@ namespace StatusERP.Services.Implementations.AS
     {
         private readonly ICobradorRepository _repository;
         private readonly ILogger<CobradorService> _logger;
-        public CobradorService(ICobradorRepository repository,ILogger<CobradorService> logger )
+        private readonly IPrivilegioUsuarioRepository _privilegioUsuarioRepository;
+
+        public CobradorService(ICobradorRepository repository, ILogger<CobradorService> logger, IPrivilegioUsuarioRepository privilegioUsuarioRepository)
         {
             _repository = repository;
             _logger = logger;
+            _privilegioUsuarioRepository = privilegioUsuarioRepository;
         }
 
-    
+        public async Task<BaseResponseGeneric<int>> CreateAsync(DtoCobrador request, string userId, string codCobrador)
 
-        public async Task<BaseResponseGeneric<int>> CreateAsync(DtoCobrador request, string userId,string codCobrador)
         {
             var response = new BaseResponseGeneric<int>();
+
             try
             {
-                var buscarCodBodega = await _repository.BuscarCodCobradorAsync(codCobrador);
-                if (buscarCodBodega != null)
+                var buscarPrivilegio = await _privilegioUsuarioRepository.GetPrivilegioUsuario("AS_COBRADOR", Constants.EmpresaId, userId);
+
+                if (buscarPrivilegio == null)
                 {
-                    throw new Exception($"El codigo de Cobrador {buscarCodBodega.CodCobrador} ya Existe");
+                    response.Errors.Add($"No tiene privilegios para crear Cobradores.");
+                    response.Success = false;
+                    return response;
+                }
+
+                var buscarCobrador = await _repository.BuscarCodCobradorAsync(codCobrador);
+                if (buscarCobrador != null)
+                {
+                    throw new Exception($"El código de Cobrador {buscarCobrador.CodCobrador} ya existe.");
                 }
                 response.Result = await _repository.CreateAsync(new Cobrador
                 {
-                     CodCobrador = request.CodCobrador,
+                    CodCobrador = request.CodCobrador,
                     Nombre = request.Nombre,
                     Email = request.Email,
-                    Activo = true,
+                    Activo = request.Activo,
+                    IsDeleted = false,
                     Updatedby = userId,
-                    CreateDate = DateTime.Now,
                     UpdateDate = DateTime.Now,
                     Createdby = userId,
+                    CreateDate = DateTime.Now
                 });
                 response.Success = true;
-
             }
             catch (Exception ex)
             {
-
                 _logger.LogCritical(ex.StackTrace);
                 response.Success = false;
                 response.Errors.Add(ex.Message);
             }
+
             return response;
         }
 
-        public async Task<BaseResponseGeneric<int>> DeleteAsync(int id,string userId)
+        public async Task<BaseResponseGeneric<int>> DeleteAsync(int id, string userId)
         {
+
             var response = new BaseResponseGeneric<int>();
             try
             {
-                await _repository.DeleteAsync(id,userId);
+                var buscarPrivilegio = await _privilegioUsuarioRepository.GetPrivilegioUsuario("AS_COBRADOR", Constants.EmpresaId, userId);
+
+                if (buscarPrivilegio == null)
+                {
+                    response.Errors.Add($"No tiene privilegios para eliminar Cobradores.");
+                    response.Success = false;
+                    return response;
+                }
+
+
+                await _repository.DeleteAsync(id, userId);
                 response.Success = true;
                 response.Result = id;
             }
@@ -72,27 +92,36 @@ namespace StatusERP.Services.Implementations.AS
                 _logger.LogCritical(ex.StackTrace);
                 response.Success = false;
                 response.Errors.Add(ex.Message);
-
             }
+
             return response;
         }
 
-        public async Task<BaseResponseGeneric<ICollection<Cobrador>>> GetAsync(int page, int rows)
+        public async Task<BaseResponseGeneric<ICollection<Cobrador>>> GetAsync(string userId)
         {
             var response = new BaseResponseGeneric<ICollection<Cobrador>>();
             try
             {
-                response.Result = await _repository.GetCollectionAsync(page, rows);
-                response.Success = true;
+                var buscarPrivilegio = await _privilegioUsuarioRepository.GetPrivilegioUsuario("AS_COBRADOR", Constants.EmpresaId, userId);
 
+                if (buscarPrivilegio == null)
+                {
+                    response.Errors.Add($"No tiene privilegios para consultar Cobradores.");
+                    response.Success = false;
+                    return response;
+                }
+
+
+                response.Result = await _repository.GetCollectionAsync();
+                response.Success = true;
             }
             catch (Exception ex)
             {
-
                 _logger.LogCritical(ex.StackTrace);
                 response.Success = false;
                 response.Errors.Add(ex.Message);
             }
+
             return response;
         }
 
@@ -106,36 +135,47 @@ namespace StatusERP.Services.Implementations.AS
             }
             catch (Exception ex)
             {
-
                 _logger.LogCritical(ex.StackTrace);
                 response.Success = false;
                 response.Errors.Add(ex.Message);
             }
+
             return response;
         }
 
-        public async Task<BaseResponseGeneric<int>> UpdateAsync(int id, DtoCobrador request ,string userId)
+        public async Task<BaseResponseGeneric<int>> UpdateAsync(int id, DtoCobrador request, string userId)
         {
             var response = new BaseResponseGeneric<int>();
             try
             {
+                var buscarPrivilegio = await _privilegioUsuarioRepository.GetPrivilegioUsuario("AS_COBRADOR", Constants.EmpresaId, userId);
+
+                if (buscarPrivilegio == null)
+                {
+                    response.Errors.Add($"No tiene privilegios para modificar Cobradores.");
+                    response.Success = false;
+                    return response;
+                }
+
+
                 response.Result = await _repository.UpdateAsync(new Cobrador
                 {
                     Id = id,
+                    CodCobrador = request.CodCobrador,
                     Nombre = request.Nombre,
                     Email = request.Email,
                     Activo = request.Activo,
-                    CodCobrador = request.CodCobrador,
-                    UpdateDate=DateTime.Now,
-                    Updatedby=userId
+                    IsDeleted = false,
+                    Updatedby = userId,
+                    UpdateDate = DateTime.Now,
                 });
+                response.Success = true;
             }
             catch (Exception ex)
             {
                 _logger.LogCritical(ex.StackTrace);
                 response.Success = false;
                 response.Errors.Add(ex.Message);
-                throw;
             }
             return response;
         }
